[flutter_tools] split application package dependencies
