+++ packages/flutter_tools/lib/src/context_runner.dart
@@ -37,6 +37,7 @@ import 'devtools_launcher.dart';
 import 'doctor.dart';
 import 'emulator.dart';
 import 'features.dart';
+import 'flutter_application_package.dart';
 import 'fuchsia/fuchsia_device.dart' show FuchsiaDeviceTools;
 import 'fuchsia/fuchsia_sdk.dart' show FuchsiaSdk, FuchsiaArtifacts;
 import 'fuchsia/fuchsia_workflow.dart' show FuchsiaWorkflow, fuchsiaWorkflow;
@@ -116,7 +117,7 @@ Future<T> runInContext<T>(
         featureFlags: featureFlags,
         operatingSystemUtils: globals.os,
       ),
-      ApplicationPackageFactory: () => ApplicationPackageFactory(
+      ApplicationPackageFactory: () => FlutterApplicationPackageFactory(
         userMessages: globals.userMessages,
         processManager: globals.processManager,
         logger: globals.logger,
