+++ packages/flutter_tools/lib/src/ios/mac.dart
@@ -7,7 +7,6 @@
 import 'package:meta/meta.dart';
 import 'package:process/process.dart';
 
-import '../application_package.dart';
 import '../artifacts.dart';
 import '../base/common.dart';
 import '../base/file_system.dart';
@@ -24,6 +23,7 @@ import '../macos/cocoapod_utils.dart';
 import '../macos/xcode.dart';
 import '../project.dart';
 import '../reporting/reporting.dart';
+import 'application_package.dart';
 import 'code_signing.dart';
 import 'devices.dart';
 import 'migrations/project_base_configuration_migration.dart';
