+++ packages/flutter_tools/lib/src/ios/simulators.dart
@@ -26,6 +26,7 @@ import '../globals.dart' as globals;
 import '../macos/xcode.dart';
 import '../project.dart';
 import '../protocol_discovery.dart';
+import 'application_package.dart';
 import 'mac.dart';
 import 'plist_parser.dart';
